requirements.txt
tk
ttkbootstrap
customtkinter
pillow
numpy
sympy
